+++ lib/hanami/assets/errors.rb
@@ -11,7 +11,7 @@ module Hanami
     end
 
     # @since 2.1.0
-    class PrecompileError < Error
+    class CompileError < Error
     end
   end
 end